Add search bars on pages, add logic for expense calculation

+++ Data/Migrations/Add_jobs_table.cs
@@ -23,7 +23,7 @@ namespace PI2022.Data.Migrations
                     PotrebnaOprema = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     PocetakRadova = table.Column<DateTime>(type: "datetime2", nullable: false),
                     ZavrsetakRadova = table.Column<DateTime>(type: "datetime2", nullable: false),
-                    Trosak = table.Column<int>(type: "int", nullable: false)
+                    Trosak = table.Column<Double>(type: "float", nullable: false)
                 },
                 constraints: table =>
                 {
