+++ Models/Jobs.cs
@@ -11,6 +11,6 @@
         public string PotrebnaOprema { get; set; }
         public DateTime PocetakRadova { get; set; }
         public DateTime ZavrsetakRadova { get; set; }
-        public int Trosak { get; set; }
+        public Double Trosak { get; set; }
     }
 }
